Added expressjs-style routing
Most changes are in router.js.
I npm installed path-to-regexp, used by the expressjs router.
I updated routes.js to work with the new router and added a route with
a path param.
I updated header.js and main.js to show an example of a route with a
path param working (/main/:id -> /main/1234)